+++ telemetry-lib/src/main/java/com/yammer/telemetry/tracing/SpanSource.java
@@ -1,10 +1,11 @@
 package com.yammer.telemetry.tracing;
 
+import java.math.BigInteger;
 import java.util.Collection;
 
 public interface SpanSource {
 
     Collection<Trace> getTraces();
 
-    Trace getTrace(long traceId);
+    Trace getTrace(BigInteger traceId);
 }
